add support for get service state for csv,text & excel files
